refactor: move assets file from helpers dir to the domain root
